Apim policy cors jwt variables

+++ src/main/java/com/nexigroup/pagopa/cruscotto/sert/service/dto/PositionPaymentDTO.java
@@ -19,7 +19,7 @@ public class PositionPaymentDTO implements Serializable {
     @JsonProperty("position-info")
     private PositionPaymentInfoDTO positionInfo;
 
-    private Integer tokens;
+    private Long tokens;
 
     @JsonProperty("all-tokens")
     private List<String> allTokens;
